gcp/requirements.txt
tensorflow==2.17.1
pillow==11.1.0
google-cloud-storage==3.1.0
keras==3.5.0